Add parsing, more verbose output
Too many things for a single commit.

Add parsing capabilities to graph.py; for now, simple directed weighted
graphs.

Add a more verbose output to the shortest_path method; if the verbose
output is selected, a table is printed with all the steps made explicit.

Add an iterator method to the queues in _queue.py.